# Fix token timing for parakeet-tdt-v2