### Why is this change needed?
<!-- Explain the motivation for this change. What problem does it solve? -->

Before:

V3: 
`[23:59:52.122] [DEBUG] [FluidAudio.Transcribe] Token timings (count: 5): [0] ' Go' (id: 4285, start: 0.640s, end: 0.960s, conf: 0.985), [1] ' a' (id: 279, start: 0.960s, end: 1.280s, conf: 0.699), [2] 'he' (id: 388, start: 1.280s, end: 1.520s, conf: 1.000), [3] 'ad' (id: 319, start: 1.520s, end: 1.920s, conf: 1.000), [4] '.' (id: 7883, start: 1.920s, end: 2.000s, conf: 0.682)`

V2
`[00:01:05.185] [DEBUG] [FluidAudio.Transcribe] Token timings (count: 6): [0] '▁G' (id: 219, start: 0.640s, end: 0.880s, conf: 0.983), [1] 'o' (id: 822, start: 0.880s, end: 1.040s, conf: 1.000), [2] '▁a' (id: 3, start: 1.040s, end: 1.280s, conf: 0.961), [3] 'he' (id: 546, start: 1.280s, end: 1.360s, conf: 1.000), [4] 'ad' (id: 103, start: 1.360s, end: 1.520s, conf: 1.000), [5] '.' (id: 841, start: 1.520s, end: 1.600s, conf: 0.951)`

After:

v3:
`[00:02:42.981] [DEBUG] [FluidAudio.Transcribe] Token timings (count: 5): [0] ' Go' (id: 4285, start: 0.640s, end: 0.960s, conf: 0.985), [1] ' a' (id: 279, start: 0.960s, end: 1.280s, conf: 0.699), [2] 'he' (id: 388, start: 1.280s, end: 1.520s, conf: 1.000), [3] 'ad' (id: 319, start: 1.520s, end: 1.920s, conf: 1.000), [4] '.' (id: 7883, start: 1.920s, end: 2.000s, conf: 0.682)`


V2:
`[00:03:33.396] [DEBUG] [FluidAudio.Transcribe] Token timings (count: 6): [0] ' G' (id: 219, start: 0.640s, end: 0.880s, conf: 0.983), [1] 'o' (id: 822, start: 0.880s, end: 1.040s, conf: 1.000), [2] ' a' (id: 3, start: 1.040s, end: 1.280s, conf: 0.961), [3] 'he' (id: 546, start: 1.280s, end: 1.360s, conf: 1.000), [4] 'ad' (id: 103, start: 1.360s, end: 1.520s, conf: 1.000), [5] '.' (id: 841, start: 1.520s, end: 1.600s, conf: 0.951)`